requirements.txt
certifi==2020.6.20
certstream==1.10
chardet==3.0.4
idna==2.10
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
termcolor==1.1.0
testresources==2.0.1
urllib3==1.25.10
websocket-client==0.48.0
watchdog==0.10.3